Print opening `describe` block
At Sencha, we almost always had nested `describe` blocks, so it was
sufficient to print the "root" `describe` block as just the test suite
name.  However, when there are not nested `describe` blocks and the
"root" is the only one, it doesn't print anything besides the suite name
(when not in verbose mode, of course), and this looks like a bug.

This commit will now always print the "root" `describe` block.

Also, bumped the version to 0.3.0